+++ CHANGELOG.md
@@ -143,6 +143,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   other slots present
 - `menu`(open) event - showed the previous menu again
 - non .yml files causing errors when loading quest packages
+- `pickrandom` event did not calculated chance correctly
 - Things that are also fixed in 1.12.X:
     - parsing of math variable
     - Citizens compatibility for not spawned NPCs